// src/app/shared/models/data.ts
export interface Ifiles {
  name: string;
  size: number;
  type: string;
}
export interface cPlayers {
  name: string;
  country: string;
  role: string;
  age: number;
  totalRuns: number,
  matchesPlayed: number;
  totalWickets: number;
}
export interface emplo {
  id: number;
  name: string;
  city: string;
  salary: number;
  dob: Date;
}
export interface Itemplo {
  name: string;
  gender: string;
  age: number;
  department: string;
  salary: number;
  yearsOfExperience: number;

}